minor updates
Split SAR_Tools.py
Added check polsartools installation
